simplify shared block creation and retention for tests
also reduce duplicate test runs
